+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerHeaderFragment.kt
@@ -56,6 +56,7 @@ import au.com.shiftyjelly.pocketcasts.ui.extensions.themed
 import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.ui.theme.ThemeColor
+import au.com.shiftyjelly.pocketcasts.utils.extensions.dpToPx
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.BookmarkFeatureControl
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
@@ -77,6 +78,7 @@ import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 private const val UP_NEXT_FLING_VELOCITY_THRESHOLD = 1000.0f
@@ -157,6 +159,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
             ShelfItem.Archive to binding.archive,
             ShelfItem.Bookmark to binding.bookmark,
             ShelfItem.Transcript to binding.transcript,
+            ShelfItem.Download to binding.download,
             ShelfItem.Report to binding.report,
         )
         viewModel.trimmedShelfLive.observe(viewLifecycleOwner) {
@@ -214,6 +217,17 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
             trackShelfAction(ShelfItem.Report.analyticsValue)
             openUrl(settings.getReportViolationUrl())
         }
+        binding.download?.setOnClickListener {
+            trackShelfAction(ShelfItem.Download.analyticsValue)
+            viewModel.handleDownloadClickFromPlaybackActions(
+                onDownloadStart = {
+                    showSnackBar(text = getString(LR.string.episode_queued_for_download))
+                },
+                onDeleteStart = {
+                    showSnackBar(text = getString(LR.string.episode_was_removed))
+                },
+            )
+        }
         binding.videoView.playbackManager = playbackManager
         binding.videoView.setOnClickListener { onFullScreenVideoClick() }
 
@@ -255,6 +269,34 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
 
             headerViewModel.episode?.let { episode ->
                 loadArtwork(episode, headerViewModel.useEpisodeArtwork, binding.artwork)
+
+                val isPodcast = episode is PodcastEpisode
+
+                val downloadIcon = when {
+                    isPodcast && (episode.isDownloading || episode.isQueued) -> IR.drawable.ic_download
+                    isPodcast && episode.isDownloaded -> IR.drawable.ic_downloaded_24dp
+                    else -> IR.drawable.ic_download
+                }
+
+                binding.download?.apply {
+                    setImageResource(downloadIcon)
+
+                    contentDescription = when {
+                        isPodcast && (episode.isDownloading || episode.isQueued) -> context.getString(LR.string.episode_downloading)
+                        isPodcast && episode.isDownloaded -> context.getString(LR.string.remove_downloaded_file)
+                        else -> context.getString(LR.string.download)
+                    }
+
+                    layoutParams = layoutParams?.apply {
+                        width = 0.dpToPx(context)
+                        height = 48.dpToPx(context)
+                    }
+
+                    scaleType = ImageView.ScaleType.CENTER
+
+                    val padding = 12.dpToPx(context)
+                    setPadding(padding, padding, padding, padding)
+                }
             }
 
             binding.podcastTitle.setOnClickListener {
@@ -678,5 +720,14 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
             .setBackgroundTint(ThemeColor.primaryUi01(Theme.ThemeType.DARK))
             .setTextColor(ThemeColor.primaryText01(Theme.ThemeType.DARK))
             .show()
+    }
+
+    private fun showSnackBar(text: CharSequence) {
+        parentFragment?.view?.let {
+            Snackbar.make(it, text, Snackbar.LENGTH_SHORT)
+                .setBackgroundTint(ThemeColor.primaryUi01(Theme.ThemeType.LIGHT))
+                .setTextColor(ThemeColor.primaryText01(Theme.ThemeType.LIGHT))
+                .show()
+        }
     }
 }