+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/SubscribeManager.kt
@@ -1,5 +1,6 @@
 package au.com.shiftyjelly.pocketcasts.repositories.podcast
 
+import android.annotation.SuppressLint
 import android.content.Context
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
 import au.com.shiftyjelly.pocketcasts.models.db.AppDatabase
@@ -61,14 +62,19 @@ class SubscribeManager @Inject constructor(
     private val episodeDao = appDatabase.episodeDao()
     private val imageRequestFactory = PocketCastsImageRequestFactory(context, isDarkTheme = true)
 
-    data class PodcastSubscribe(val podcastUuid: String, val sync: Boolean)
+    data class PodcastSubscribe(val podcastUuid: String, val sync: Boolean, val shouldAutoDownload: Boolean)
 
+    @SuppressLint("CheckResult")
     private fun setupSubscribeRelay(): PublishRelay<PodcastSubscribe> {
         val source = PublishRelay.create<PodcastSubscribe>()
         source
             .observeOn(Schedulers.io())
             .doOnNext { info -> Timber.i("Adding podcast to addPodcast queue ${info.podcastUuid}") }
-            .flatMap({ info -> addPodcast(info.podcastUuid, sync = info.sync, subscribed = true).toObservable() }, true, 5)
+            .flatMap({ info ->
+                // shouldAutoDownload = true because the user manually subscribed to the podcast,
+                // so we want to automatically download episodes at this moment.
+                addPodcast(info.podcastUuid, sync = info.sync, subscribed = true, shouldAutoDownload = info.shouldAutoDownload).toObservable()
+            }, true, 5)
             .doOnError { throwable -> LogBuffer.e(LogBuffer.TAG_BACKGROUND_TASKS, throwable, "Could not subscribe to podcast") }
             .subscribeBy(
                 onNext = { podcast ->
@@ -83,19 +89,19 @@ class SubscribeManager @Inject constructor(
     /**
      * Subscribe to a podcast on a background queue.
      */
-    fun subscribeOnQueue(podcastUuid: String, sync: Boolean = false) {
+    fun subscribeOnQueue(podcastUuid: String, sync: Boolean = false, shouldAutoDownload: Boolean) {
         if (uuidsInQueue.contains(podcastUuid)) {
             return
         }
         uuidsInQueue.add(podcastUuid)
-        subscribeRelay.accept(PodcastSubscribe(podcastUuid, sync))
+        subscribeRelay.accept(PodcastSubscribe(podcastUuid, sync, shouldAutoDownload))
     }
 
     /**
      * Subscribe to a podcast and wait.
      */
-    fun addPodcast(podcastUuid: String, sync: Boolean = false, subscribed: Boolean = false): Single<Podcast> {
-        return subscribeToExistingOrServerPodcast(podcastUuid, sync, subscribed)
+    fun addPodcast(podcastUuid: String, sync: Boolean = false, subscribed: Boolean = false, shouldAutoDownload: Boolean): Single<Podcast> {
+        return subscribeToExistingOrServerPodcast(podcastUuid, sync, subscribed, shouldAutoDownload)
             .flatMap {
                 LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "Adding podcast $podcastUuid to database")
                 cacheArtwork(it).toSingleDefault(it)
@@ -105,15 +111,17 @@ class SubscribeManager @Inject constructor(
                 // update the notification time as any podcasts added after this date will be ignored
                 settings.setNotificationLastSeenToNow()
 
-                if (subscribed && FeatureFlag.isEnabled(Feature.AUTO_DOWNLOAD)) {
+                if (canDownloadEpisodesAfterSubscription(subscribed, shouldAutoDownload)) {
                     podcastDao.findByUuid(podcastUuid)?.let { podcast ->
                         val episodes = episodeManager.findEpisodesByPodcastOrderedByPublishDate(podcast)
-                        val autoDownloadLimit = podcast.autoDownloadLimit ?: settings.autoDownloadLimit.value
+                        val numberOfEpisodes = AutoDownloadLimitSetting.getNumberOfEpisodes(settings.autoDownloadLimit.value)
 
-                        episodes.take(AutoDownloadLimitSetting.getNumberOfEpisodes(autoDownloadLimit)).forEach { episode ->
+                        episodes.take(numberOfEpisodes).forEach { episode ->
                             if (episode.isQueued || episode.isDownloaded || episode.isDownloading || episode.isExemptFromAutoDownload) {
                                 return@forEach
                             }
+
+                            LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "Auto Downloading $numberOfEpisodes episodes after subscribing to $podcastUuid")
 
                             DownloadHelper.addAutoDownloadedEpisodeToQueue(
                                 episode,
@@ -146,7 +154,7 @@ class SubscribeManager @Inject constructor(
         return uuidsInQueue
     }
 
-    private fun subscribeToExistingOrServerPodcast(podcastUuid: String, sync: Boolean, subscribed: Boolean): Single<Podcast> {
+    private fun subscribeToExistingOrServerPodcast(podcastUuid: String, sync: Boolean, subscribed: Boolean, shouldAutoDownload: Boolean): Single<Podcast> {
         // check if the podcast exists already
         val subscribedObservable = podcastDao.isSubscribedToPodcastRx(podcastUuid)
         return subscribedObservable.flatMap { isSubscribed ->
@@ -154,7 +162,7 @@ class SubscribeManager @Inject constructor(
             if (isSubscribed) {
                 subscribeToExistingPodcast(podcastUuid, sync)
             } else {
-                subscribeToServerPodcast(podcastUuid, sync, subscribed)
+                subscribeToServerPodcast(podcastUuid, sync, subscribed, shouldAutoDownload)
             }
         }
     }
@@ -170,7 +178,7 @@ class SubscribeManager @Inject constructor(
         return updateObservable.andThen(findObservable.toSingle())
     }
 
-    private fun subscribeToServerPodcast(podcastUuid: String, sync: Boolean, subscribed: Boolean): Single<Podcast> {
+    private fun subscribeToServerPodcast(podcastUuid: String, sync: Boolean, subscribed: Boolean, shouldAutoDownload: Boolean): Single<Podcast> {
         // download the podcast
         val podcastObservable = downloadPodcast(podcastUuid)
             .doOnSuccess { podcast ->
@@ -179,9 +187,9 @@ class SubscribeManager @Inject constructor(
                 podcast.isSubscribed = subscribed
                 podcast.grouping = settings.podcastGroupingDefault.value
                 podcast.showArchived = settings.showArchivedDefault.value
-                if (subscribed && FeatureFlag.isEnabled(Feature.AUTO_DOWNLOAD)) {
+                if (canDownloadEpisodesAfterSubscription(subscribed, shouldAutoDownload)) {
+                    LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "Update auto download status for $podcastUuid")
                     podcast.autoDownloadStatus = AUTO_DOWNLOAD_NEW_EPISODES
-                    podcast.autoDownloadLimit = settings.autoDownloadLimit.value
                 }
             }
         // add the podcast
@@ -191,6 +199,9 @@ class SubscribeManager @Inject constructor(
         // insert episodes
         return insertPodcastObservable.flatMap { podcast -> subscribeInsertEpisodes(podcast).toSingle { podcast } }
     }
+
+    private fun canDownloadEpisodesAfterSubscription(subscribed: Boolean, shouldAutoDownload: Boolean) =
+        subscribed && settings.autoDownloadNewEpisodes.value && shouldAutoDownload && FeatureFlag.isEnabled(Feature.AUTO_DOWNLOAD)
 
     private fun downloadPodcast(podcastUuid: String): Single<Podcast> {
         // download the podcast